+++ src/customize/core/__tests__/MockKintoneApiClient.ts
@@ -1,8 +1,8 @@
 import type {
   Option as ApiClientOption,
   UpdateAppCustomizeParameter,
-} from "../../KintoneApiClient";
-import KintoneApiClient from "../../KintoneApiClient";
+} from "../client";
+import KintoneApiClient from "../client";
 
 type MethodParameter = "GET" | "POST" | "PUT" | "DELETE";
