+++ src/customize/init/index.ts
@@ -1,8 +1,7 @@
 import fs from "fs";
 import { mkdirp } from "mkdirp";
-import type { CustomizeManifest } from "./index";
-import type { Lang } from "../lang";
-import { getBoundMessage } from "../messages";
+import { getBoundMessage } from "../core";
+import type { CustomizeManifest, Lang } from "../core";
 
 export const getInitCustomizeManifest = (
   appId: string,
