+++ src/customize/init/params.ts
@@ -1,6 +1,6 @@
-import type { Lang } from "../lang";
-import { getBoundMessage } from "../messages";
-import { promptForAppId, promptForScope } from "./prompts/init";
+import { getBoundMessage } from "../core";
+import type { Lang } from "../core";
+import { promptForAppId, promptForScope } from "./prompts";
 
 export const inquireInitParams = async (lang: Lang) => {
   const m = getBoundMessage(lang);
